Got express app module working
